+++ src/stores/root-store.ts
@@ -22,6 +22,7 @@ export class RootStore {
 export type RootStoreHydration = Partial<{
   room: RoomStatus;
   refreshSeconds: number;
+  reservationEnabled: boolean;
 }>;
 
 let browserStore: RootStore | undefined;
@@ -29,7 +30,11 @@ let browserStore: RootStore | undefined;
 export function createRootStore(initialData?: RootStoreHydration): RootStore {
   const store = new RootStore();
   if (initialData?.room) {
-    store.machines.hydrate(initialData.room, initialData.refreshSeconds);
+    store.machines.hydrate(
+      initialData.room,
+      initialData.refreshSeconds,
+      initialData.reservationEnabled,
+    );
   }
   return store;
 }